+++ pysatMissions/tests/test_instruments.py
@@ -23,8 +23,6 @@ from pysat.tests.instrument_test_class import InstTestClass
 from pysat.utils import generate_instrument_list
 
 
-saved_path = pysat.params['data_dirs']
-
 # Developers for instrument libraries should update the following line to
 # point to their own subpackage location
 # e.g.,
@@ -89,6 +87,47 @@ class TestInstruments(InstTestClass):
         return
 
     # Custom package unit tests can be added here
+
+    @pytest.mark.parametrize("kwargs",
+                             [{},
+                              {'inclination': 20, 'alt_periapsis': 400},
+                              {'inclination': 80, 'alt_periapsis': 500,
+                               'alt_apoapsis': 600,
+                               'epoch': dt.datetime(2019, 1, 1)}])
+    def test_sgp4_data_continuity(self, kwargs):
+        """Test that data is continuous for sequential days.
+
+        Parameters
+        ----------
+        kwargs : dict
+            Optional kwargs to pass through.  If empty, instrument will be
+            default TLEs.
+
+        """
+
+        # Define sat with custom Keplerian inputs
+        sat = pysat.Instrument(
+            inst_module=pysatMissions.instruments.missions_sgp4,
+            **kwargs)
+
+        # Get last 10 points of day 1
+        sat.load(2018, 1)
+        day1 = sat.data[-10:]
+
+        # Get first 10 points of day 2
+        sat.load(2018, 2)
+        day2 = sat.data[:10]
+
+        average_gradient = day1.diff().mean()
+        std_gradient = day1.diff().std()
+        gradient_between_days = day2.iloc[0] - day1.iloc[-1]
+
+        # Check that the jump between days is within 3 sigma of average gradient
+        del_g = np.abs(average_gradient - gradient_between_days)
+        assert np.all(del_g < (3. * std_gradient)), \
+            "Gap between days outside of 3 sigma"
+
+        return
 
     @pytest.mark.parametrize("inst_dict", [x for x in instruments['download']])
     @pytest.mark.parametrize("kwarg,output", [(None, 1), ('10s', 10)])