+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Update sgp4 interface to use new syntax for initialization from TLEs
 * Include conversions to geodetic latitude / longitude / altitude for sgp4
 * Improve metadata generation in missions_sgp4
+* Update syntax to be compliant with OMMBV 1.0
 * Documentation
   * Improve docstrings throughout
 * Testing