+++ CHANGELOG.md
@@ -7,10 +7,12 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Update sgp4 interface to use new syntax for initialization from TLEs
 * Include conversions to geodetic latitude / longitude / altitude for sgp4
 * Improve metadata generation in missions_sgp4
+* Update syntax to be compliant with OMMBV 1.0
 * Documentation
   * Improve docstrings throughout
 * Testing
   * Add style check for docstrings
+  * Improve checks in codeclimate
 
 ## [0.2.2] - 2021-06-18
 * Migrate pyglow interface to pysatIncubator