+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Update sgp4 interface to use new syntax for initialization from TLEs
 * Include conversions to geodetic latitude / longitude / altitude for sgp4
 * Improve metadata generation in missions_sgp4
+* Update syntax to be compliant with OMMBV 1.0
 * Documentation
   * Improve docstrings throughout
 * Deprecations
@@ -14,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Testing
   * Add style check for docstrings
   * Added checks for deprecation warnings
+  * Improve checks in codeclimate
 
 ## [0.2.2] - 2021-06-18
 * Migrate pyglow interface to pysatIncubator